// Backend/src/constant/enum.ts
export enum Role {
  CANDIDATE = "candidate",
  RECRUITER = "recruiter",
  ADMIN = "admin",
}

export enum JobType {
  FULL_TIME = "full-time",
  PART_TIME = "part-time",
  CONTRACT = "contract",
  FREELANCE = "freelance",
  INTERNSHIP = "internship",
}

export enum WorkMode {
  REMOTE = "remote",
  ONSITE = "onsite",
  HYBRID = "hybrid",
}

export enum ApplicationStatus {
  PENDING = "pending",
  REVIEWING = "reviewing",
  INTERVIEWING = "interviewing",
  REJECTED = "rejected",
  HIRED = "hired",
}

export enum JobStatus {
  DRAFT = "draft",
  PUBLISHED = "published",
  CLOSED = "closed",
}